+++ pgpy/constants.py
@@ -12,18 +12,15 @@ from enum import IntEnum
 from enum import IntFlag
 
 from types import LambdaType
-from typing import Optional
+from typing import Optional, NamedTuple, Type, Union
 
-from pyasn1.type.univ import ObjectIdentifier
-
-from cryptography.hazmat.backends import openssl
-from cryptography.hazmat.primitives.asymmetric import ec
+from cryptography.hazmat.primitives.asymmetric import ec, x25519, ed25519
 from cryptography.hazmat.primitives.ciphers import algorithms
 
 from .decorators import classproperty
-from ._curves import BrainpoolP256R1, BrainpoolP384R1, BrainpoolP512R1, X25519, Ed25519
 
 __all__ = [
+    'ECFields',
     'EllipticCurveOID',
     'ECPointFormat',
     'PacketTag',
@@ -220,88 +217,132 @@ class SymmetricKeyAlgorithm(IntEnum):
         return os.urandom(self.key_size // 8)
 
 
-class EllipticCurveOID(Enum):
-    """OIDs for supported elliptic curves."""
-    # these are specified as:
-    # id = (oid, curve)
-    Invalid = ('', )
-    #: DJB's fast elliptic curve
-    Curve25519 = ('1.3.6.1.4.1.3029.1.5.1', X25519)
-    #: Twisted Edwards variant of Curve25519
-    Ed25519 = ('1.3.6.1.4.1.11591.15.1', Ed25519)
-    #: NIST P-256, also known as SECG curve secp256r1
-    NIST_P256 = ('1.2.840.10045.3.1.7', ec.SECP256R1)
-    #: NIST P-384, also known as SECG curve secp384r1
-    NIST_P384 = ('1.3.132.0.34', ec.SECP384R1)
-    #: NIST P-521, also known as SECG curve secp521r1
-    NIST_P521 = ('1.3.132.0.35', ec.SECP521R1)
-    #: Brainpool Standard Curve, 256-bit
-    #:
-    #: .. note::
-    #:     Requires OpenSSL >= 1.0.2
-    Brainpool_P256 = ('1.3.36.3.3.2.8.1.1.7', BrainpoolP256R1)
-    #: Brainpool Standard Curve, 384-bit
-    #:
-    #: .. note::
-    #:     Requires OpenSSL >= 1.0.2
-    Brainpool_P384 = ('1.3.36.3.3.2.8.1.1.11', BrainpoolP384R1)
-    #: Brainpool Standard Curve, 512-bit
-    #:
-    #: .. note::
-    #:     Requires OpenSSL >= 1.0.2
-    Brainpool_P512 = ('1.3.36.3.3.2.8.1.1.13', BrainpoolP512R1)
-    #: SECG curve secp256k1
-    SECP256K1 = ('1.3.132.0.10', ec.SECP256K1)
+class ECFields(NamedTuple):
+    name:str
+    OID:str
+    OID_der:bytes
+    key_size:int # in bits
+    kdf_halg:HashAlgorithm
+    kek_alg:SymmetricKeyAlgorithm
+    curve:Type
 
-    def __new__(cls, oid, curve=None) -> "EllipticCurveOID":
+    def __repr__(self) -> str:
+        return f'<Elliptic Curve {self.name} ({self.OID})>'
+
+class EllipticCurveOID(Enum):
+    """Supported elliptic curves."""
+
+    #: DJB's fast elliptic curve
+    Curve25519 = (x25519, '1.3.6.1.4.1.3029.1.5.1',
+                  b'\x2b\x06\x01\x04\x01\x97\x55\x01\x05\x01',
+                  'X25519', 256)
+    #: Twisted Edwards variant of Curve25519
+    Ed25519 = (ed25519, '1.3.6.1.4.1.11591.15.1',
+               b'\x2b\x06\x01\x04\x01\xda\x47\x0f\x01',
+               'Ed25519', 256)
+    #: NIST P-256, also known as SECG curve secp256r1
+    NIST_P256 = (ec.SECP256R1, '1.2.840.10045.3.1.7',
+                b'\x2a\x86\x48\xce\x3d\x03\x01\x07')
+    #: NIST P-384, also known as SECG curve secp384r1
+    NIST_P384 = (ec.SECP384R1, '1.3.132.0.34',
+                 b'\x2b\x81\x04\x00\x22')
+    #: NIST P-521, also known as SECG curve secp521r1
+    NIST_P521 = (ec.SECP521R1, '1.3.132.0.35',
+                 b'\x2b\x81\x04\x00\x23')
+    #: Brainpool Standard Curve, 256-bit
+    Brainpool_P256 = (ec.BrainpoolP256R1, '1.3.36.3.3.2.8.1.1.7',
+                      b'\x2b\x24\x03\x03\x02\x08\x01\x01\x07')
+    #: Brainpool Standard Curve, 384-bit
+    Brainpool_P384 = (ec.BrainpoolP384R1, '1.3.36.3.3.2.8.1.1.11',
+                      b'\x2b\x24\x03\x03\x02\x08\x01\x01\x0b')
+    #: Brainpool Standard Curve, 512-bit
+    Brainpool_P512 = (ec.BrainpoolP512R1, '1.3.36.3.3.2.8.1.1.13',
+                      b'\x2b\x24\x03\x03\x02\x08\x01\x01\x0d')
+    #: SECG curve secp256k1
+    SECP256K1 = (ec.SECP256K1, '1.3.132.0.10',
+                 b'\x2b\x81\x04\x00\x0a')
+
+    def __new__(cls, impl_cls:Type, oid:str, oid_der:bytes, name:Optional[str]=None, key_size_bits:Optional[int]=None) -> "EllipticCurveOID":
         # preprocessing stage for enum members:
         #  - set enum_member.value to ObjectIdentifier(oid)
         #  - if curve is not None and curve.name is in ec._CURVE_TYPES, set enum_member.curve to curve
         #  - otherwise, set enum_member.curve to None
         obj = object.__new__(cls)
-        obj._value_ = ObjectIdentifier(oid)
-        obj.curve = None
+        if name is None:
+            newname = impl_cls.name
+            if not isinstance(newname, str):
+                raise TypeError(f"{impl_cls}.name is not string!")
+            name = newname
+        if key_size_bits is None:
+            newks = impl_cls.key_size
+            if not isinstance(newks, int):
+                raise TypeError(f"{impl_cls}.name is not string!")
+            key_size_bits = newks
 
-        if curve is not None and curve.name in ec._CURVE_TYPES:
-            obj.curve = curve
+        algs = {256: (HashAlgorithm.SHA256, SymmetricKeyAlgorithm.AES128),
+                384: (HashAlgorithm.SHA384, SymmetricKeyAlgorithm.AES192),
+                512: (HashAlgorithm.SHA512, SymmetricKeyAlgorithm.AES256),
+                521: (HashAlgorithm.SHA512, SymmetricKeyAlgorithm.AES256)}
+
+        (kdf_alg, kek_alg) = algs[key_size_bits]
+
+        obj._value_ = ECFields(name, oid, oid_der, key_size_bits, kdf_alg, kek_alg, impl_cls)
 
         return obj
 
-    @property
-    def can_gen(self) -> bool:
-        return self.curve is not None
-
-    @property
-    def key_size(self) -> Optional[int]:
-        if self.curve is not None:
-            return self.curve.key_size
+    @classmethod
+    def from_key_size(cls, key_size:int) -> Optional["EllipticCurveOID"]:
+        for c in EllipticCurveOID:
+            if c.value.key_size == key_size:
+                return c
+        warnings.warn(f"Cannot find any Elliptic curve of size: {key_size}")
         return None
 
+    @classmethod
+    def from_OID(cls, oid:bytes) -> Union["EllipticCurveOID", bytes]:
+        for c in EllipticCurveOID:
+            if c.value.OID_der == oid:
+                return c
+        warnings.warn(f"Unknown Elliptic curve OID: {oid!r}")
+        return oid
+
+    @classmethod
+    def parse(cls, packet:bytearray) -> Union["EllipticCurveOID", bytes]:
+        oidlen = packet[0]
+        del packet[0]
+        ret = EllipticCurveOID.from_OID(bytes(packet[:oidlen]))
+        del packet[:oidlen]
+        return ret
+
     @property
-    def kdf_halg(self) -> Optional[HashAlgorithm]:
-        # return the hash algorithm to specify in the KDF fields when generating a key
-        algs = {256: HashAlgorithm.SHA256,
-                384: HashAlgorithm.SHA384,
-                512: HashAlgorithm.SHA512,
-                521: HashAlgorithm.SHA512}
-
-        ksz = self.key_size
-        if ksz is None:
-            return None
-        return algs.get(ksz, None)
+    def key_size(self) -> int:
+        return self.value.key_size
 
     @property
-    def kek_alg(self) -> Optional[SymmetricKeyAlgorithm]:
-        # return the AES algorithm to specify in the KDF fields when generating a key
-        algs = {256: SymmetricKeyAlgorithm.AES128,
-                384: SymmetricKeyAlgorithm.AES192,
-                512: SymmetricKeyAlgorithm.AES256,
-                521: SymmetricKeyAlgorithm.AES256}
+    def oid(self) -> str:
+        return self.value.OID
 
-        ksz = self.key_size
-        if ksz is None:
-            return None
-        return algs.get(ksz, None)
+    @property
+    def kdf_halg(self) -> HashAlgorithm:
+        return self.value.kdf_halg
+
+    @property
+    def kek_alg(self) -> SymmetricKeyAlgorithm:
+        return self.value.kek_alg
+
+    @property
+    def curve(self) -> Type:
+        return self.value.curve
+
+    @property
+    def can_gen(self) -> bool:
+        return True
+
+    def __bytes__(self) -> bytes:
+        return bytes([len(self.value.OID_der)]) + self.value.OID_der
+
+    def __len__(self) -> int:
+        return len(self.value.OID_der) + 1
 
 
 class ECPointFormat(IntEnum):