+++ pgpy/pgp.py
@@ -22,6 +22,7 @@ from typing import Optional, Tuple, Set, Union, List, Iterator, Deque
 from cryptography.hazmat.primitives import hashes
 
 from .constants import CompressionAlgorithm
+from .constants import EllipticCurveOID
 from .constants import Features
 from .constants import HashAlgorithm
 from .constants import ImageEncoding
@@ -1487,12 +1488,14 @@ class PGPKey(Armorable, ParentRef, PGPObject):
         return self._key.pkalg
 
     @property
-    def key_size(self):
+    def key_size(self) -> Optional[Union[int,EllipticCurveOID]]:
         """
         The size pertaining to this key. ``int`` for non-EC key algorithms; :py:obj:`constants.EllipticCurveOID` for EC keys.
 
         .. versionadded:: 0.4.1
         """
+        if self._key is None:
+            return None
         if self.key_algorithm in {PubKeyAlgorithm.ECDSA, PubKeyAlgorithm.ECDH, PubKeyAlgorithm.EdDSA}:
             return self._key.keymaterial.oid
         # check if keymaterial is not an Opaque class containing a bytearray